+++ packages/orq-rc/src/sdk/feedback.ts
@@ -45,10 +45,10 @@ export class Feedback extends ClientSDK {
   }
 
   /**
-   * Create chat completion
+   * Upload a file
    *
    * @remarks
-   * Creates a model response for the given chat conversation with support for retries, fallbacks, prompts, and variables.
+   * Files are used to upload documents that can be used with features like Deployments.
    */
   async create(
     request?: operations.PostV2FeedbackRequestBody | undefined,